[LeftPanel] move "Open" & "Settings" button to left panel
