+++ src/Viewer/services/clpWorker.js
@@ -83,7 +83,6 @@ onmessage = function (e) {
                 sendError(e);
             }
             break;
-
         default:
             break;
     }
